sdcard_image-rpi: Align SD partition and optimize final SD image
SD image file was 4G in size but actually the rootfs partition was only
60MB (ex.: rpi-basic-image).
The image file size is not fixed anymore but calculated inside this
class. In this way we will setup the partition for rootfs to fit the
ext file used as SDIMG_ROOTFS. By default the free space is calculated
with IMAGE_OVERHEAD_FACTOR which is by default 1.3. If more free space
is needed this can be overwritten.
IMAGE_ROOTFS_ALIGNMENT set at 4096. This should be the best value for a
SD card.

[GITHUB #20]
[GITHUB #59]

Signed-off-by: Andrei Gherzan <andrei@gherzan.ro>