linux-raspberrypi: Change to Linux Yocto style kernel recipes
Yocto style recipes provide support for application of kernel
configuration fragments which is achieved with this change.

For further details see the Yocto Linux Kernel Development Manual

ref: http://www.yoctoproject.org/docs/1.6/kernel-dev/kernel-dev.html

In addition .bb files are simplified to take a SRCBRANCH which is
used by the SRC_URI set in linux-raspberrypi.inc and together with
the existing SRCREV define a checkout from the linux-raspberrypi
git repository.

A default configuration, defconfig, is provided as a baseline.
This is generated from bcmrpi_defconfig, as used in the existing
recipes, and so configuration options should be the same.

To change the kernel configuration a new configuration fragment
may be generated, as explained in the kernel development documentation
above, and this file then added to the SRC_URI via a .bbappend.

Added some minor refactoring and removed some useless code.

Change-Id: Ic72e14ec14ec3f042aeda0c6820f896be9c02e69
Signed-off-by: Alex J Lennon <ajlennon@dynamicdevices.co.uk>
Signed-off-by: Andrei Gherzan <andrei@gherzan.ro>