docs: Initial split of README into docs directory
The top-level README document has got pretty unwieldy and so we need to
switch to add some more structure to our documentation. The first step
is to split out sub-sections of the README document into separate files
in a new 'docs' directory.

Whilst splitting up the README, we can also take the opportunity to tidy
things up and fix a couple of typos.

Signed-off-by: Paul Barker <pbarker@toganlabs.com>