base-files: add 'i2c_dev' to auto loaded modules
the Kernel module i2c_dev is required in order to be able
to successfully interact with the i2c buses on the Raspberry's
expansion pins.
Putting this module in modules-load.d, we ensure that the i2c
subsystem is properly loaded, so whenever the i2c buses are
enabled in config.txt, udev will be able to create the
necessary device nodes in /dev.

This very same file is also present on Raspian official distro.

Signed-off-by: Francesco Giancane <francescogiancane8@gmail.com>