Fix KERNEL vars having no effect on kernel recipes.
In 4d3a148b8eaca, a new default to KERNEL_IMAGETYPE_UBOOT was added to
raspberrypi3-64.conf, which means that the generic default had to be
changed from '?=' to '??='. However, this presents a problem when
building the kernel recipes, because the '??=' assignment is no longer
strong enough to win over the '?=' which exists in the upstream
kernel.bbclass file. So instead, turn the assignment back to '?=', and
rely on ordering of include files to get correct assignments for
raspberrypi3-64.

connected to #153.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>