udev-udisks-rules-rpi: Hide initial boot partition from udisks2
If I am not mistaken, many desktops (kde/lxqt/xfce..) use udisks2 to ask for
devices which can be mounted. Without this patch the initial boot partition
can be mounted by a single click on folder displayed on desktop or file manager.

Why it is suggested to add this recipe to an image:
* It removes annoying icon in desktop/file manager.
* The initial boot partition is vfat. Unpriviledged users can modify
  content. E.g by removing files accidently, images can be turned into
  non-booting condition easily. And from security point of view, it is a
  disaster.

FWIW: Have similar in all my BSPs vor very long time.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>