bluez5: fix issues with systemd script
Type=simple is wrong: it allows bluetooth.service to start before the
uart is configured, resulting in hci0 command tx timeout errors.
Type=oneshot blocks follow-up units until the ExecStart completes.

Add RemainAfterExit since system state has changed as a result of the
unit.

Also add a BindsTo for the device that we're going to use.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>