linux-firmware-rpidistro: Fix WiFi on RaspberryPi 4
The official Raspbian with RPi4 support has an additional change for the
NVRAM WiFi configuration: boardflags3 value is changed from "0x48200100"
to "0x44200100". Without this change the interface never comes up. It is
confusing because the deb changelog doesn't say anything about it and
the repository upstream doesn't have this change. Asked for more info
here:
https://github.com/RPi-Distro/firmware-nonfree/issues/3

In the meanwhile we import this change. Also, we remove the .pc directory
which confuses the patch step.

Signed-off-by: Andrei Gherzan <andrei@gherzan.ro>