lirc_%.bbappend: Remove
Temporary remove the LIRC patch because it is not compatible with
version 0.9.4d which is at the moment provided by branch dunfell
from meta-openembedded/meta-oe. This patch should be applied again
when LIRC version 0.10.1 or newer is available at branch dunfell
of meta-openembedded/meta-oe (as of the moment this version is
available at master).

Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>