bootfiles: Use BOOTFILES_DIR_NAME variable
This variable is set in rpi-base.inc and referenced by other recipes
which need to know the location of the bootfiles. If this is overridden
for any reason we need to ensure that the bootfiles recipe deploys to
the correct path.

Signed-off-by: Paul Barker <pbarker@konsulko.com>
(cherry picked from commit b8440b3d26fc0b59acd48ed6f3e39e00cd867f6c)
Signed-off-by: Ricardo Salveti <ricardo@foundries.io>