conf/machine/include/rpi-base.inc: Added a line to RPI_KERNEL_DEVICETREE_OVERLAYS to make the gpio-poweroff overlay available.
Certain systems require a GPIO to be toggled by the kernel when they are powered on and off.  This feature is enabled by a kernel overlay on the raspberry pi.  This overlay was available in raspberrypi-os but had not been added to the list of kernel overlays that are installed on yocto hardknott.

Signed-off-by: Cameron Kellough <cameron@telemetrak.com>