rpi-config: Add RPI_CONFIG_STRIP to strip down config.txt
The size of the file config.txt is over 52k, and it may grows over the
versions of the firmware. Even the deprecated options are maintained.

This adds the variable RPI_CONFIG_STRIP to allow the strip down of the
file config.txt to the very bare minimal: it removes the leading white
spaces, the empty lines and the comments.

Signed-off-by: Gilles Doffe <gilles.doffe@rtone.fr>
Signed-off-by: Gaël PORTAY <gael.portay@rtone.fr>