rpi-config: add config file template
Relying on a commented out config file fetched from a user Github
repository is likely error prone and misleading. The best practice
is to direct the user to the official documentation describing the
complete list of supported configuration keys.

Only a few keys are added to the config.txt file anyway, so start
from an empty file that the user may override if necessary.

Move the config.txt tweak code into do_compile for clarity and to
ensure that the 80 character wide check is done once the file is
modified, not in between appended modifications.

Signed-off-by: Vivien Didelot <vivien.didelot@gmail.com>