docs, conf/machine/include/rpi-base.inc, rpi-config: added seeed-can-fd-hat-v2.dtbo
I needed support of the seeed-can-fd-hat-v2 Can Shield

Added the seeed-can-fd-hat-v2.dtbo device tree overlay to the rpi-base.inc and added the parameter for the CAN shield under the rpi-config so it can be easily enabled as a parameter from e.g. the local.conf and is added to the /boot/config.txt of the Pi. I also added the changes to the documentation.

Signed-off-by: Moritz Boesenberg <moritz@techmountain.dev>