rpi-config: ENABLE_UART for CM5 IO Board
The Raspberry Pi Compute Module 5 IO Board does not include a
dedicated UART debug connector like the Raspberry Pi 5. Although
the CM5 itself has reserved space for a JST connector, it is not
populated. To enable UART on the 40-pin header of the CM 5 IO Board
we need to add dtoverlay=uart0 and dtparam=uart0_console to
config.txt.

This work was sponsored by GOVCERT.LU.

Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>